+++ media/audio/simple_sources.cc
@@ -1,6 +1,8 @@
-// Copyright (c) 2006-2008 The Chromium Authors. All rights reserved.
+// Copyright (c) 2010 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
+
+#include "media/audio/simple_sources.h"
 
 #include <algorithm>
 #include <math.h>
@@ -8,7 +10,6 @@
 #include "base/basictypes.h"
 #include "base/logging.h"
 #include "media/audio/audio_output.h"
-#include "media/audio/simple_sources.h"
 
 //////////////////////////////////////////////////////////////////////////////
 // SineWaveAudioSource implementation.
@@ -25,14 +26,14 @@ SineWaveAudioSource::SineWaveAudioSource(Format format, int channels,
 
 // The implementation could be more efficient if a lookup table is constructed
 // but it is efficient enough for our simple needs.
-size_t SineWaveAudioSource::OnMoreData(AudioOutputStream* stream, void* dest,
-                                       size_t max_size, int pending_bytes) {
+uint32 SineWaveAudioSource::OnMoreData(AudioOutputStream* stream, void* dest,
+                                       uint32 max_size, uint32 pending_bytes) {
   const double kTwoPi = 2.0 * 3.141592653589;
   double f = freq_ / sample_freq_;
   int16* sin_tbl = reinterpret_cast<int16*>(dest);
-  size_t len = max_size / sizeof(int16);
+  uint32 len = max_size / sizeof(int16);
   // The table is filled with s(t) = 32768*sin(2PI*f*t).
-  for (size_t ix = 0; ix != len; ++ix) {
+  for (uint32 ix = 0; ix != len; ++ix) {
     double th = kTwoPi * ix * f;
     sin_tbl[ix] = static_cast<int16>((1 << 15) * sin(th));
   }
@@ -49,7 +50,7 @@ void SineWaveAudioSource::OnError(AudioOutputStream* stream, int code) {
 //////////////////////////////////////////////////////////////////////////////
 // PushSource implementation.
 
-PushSource::PushSource(size_t packet_size)
+PushSource::PushSource(uint32 packet_size)
     : packet_size_(packet_size),
       buffered_bytes_(0),
       front_buffer_consumed_(0) {
@@ -59,9 +60,9 @@ PushSource::~PushSource() {
   CleanUp();
 }
 
-size_t PushSource::OnMoreData(AudioOutputStream* stream, void* dest,
-                              size_t max_size, int pending_bytes) {
-  size_t copied = 0;
+uint32 PushSource::OnMoreData(AudioOutputStream* stream, void* dest,
+                              uint32 max_size, uint32 pending_bytes) {
+  uint32 copied = 0;
   while (copied < max_size) {
     AutoLock auto_lock(lock_);
 
@@ -69,7 +70,7 @@ size_t PushSource::OnMoreData(AudioOutputStream* stream, void* dest,
     if (!packets_.size())
       break;
     Packet packet = packets_.front();
-    size_t size = std::min(max_size - copied,
+    uint32 size = std::min(max_size - copied,
                            packet.size - front_buffer_consumed_);
     memcpy(static_cast<char*>(dest) + copied,
            packet.buffer + front_buffer_consumed_,
@@ -95,7 +96,7 @@ void PushSource::OnError(AudioOutputStream* stream, int code) {
 }
 
 // TODO(cpu): Manage arbitrary large sizes.
-bool PushSource::Write(const void *data, size_t len) {
+bool PushSource::Write(const void *data, uint32 len) {
   if (len == 0) {
     NOTREACHED();
     return false;
@@ -109,7 +110,7 @@ bool PushSource::Write(const void *data, size_t len) {
   return true;
 }
 
-size_t PushSource::UnProcessedBytes() {
+uint32 PushSource::UnProcessedBytes() {
   AutoLock auto_lock(lock_);
   return buffered_bytes_;
 }
